+++ pytest.ini
@@ -1,4 +1,4 @@
 [pytest]
 DJANGO_SETTINGS_MODULE=doc_search.mysite.test_settings
-python_files = tests.py test_*.py *_tests.py
+python_files = tests.py *_test.py *_tests.py
 adopts = --nomigrations --cov=. --cov-report=html
